src/SnippetsService/requirements.txt
flask
flask-cors
pycrypto
pycryptodome
pyjwt
gunicorn
flasgger
mysql-connector-python